+++ app/src/main/java/com/openclassrooms/realestatemanager/datas/model/TypeOfProperty.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity
 data class TypeOfProperty (
     @PrimaryKey(autoGenerate = false)
-    var idType: Long,
+    var idType: Int,
     var nameType: String,
 
     ) {
